ostd

Module arch

Platform-specific code for the x86 platform.

Modules

boot

The x86 boot module defines the entrypoints of Asterinas and the corresponding headers for different x86 boot protocols.

device

Device-related APIs. This module mainly contains the APIs that should exposed to the device driver like PCI, RTC

kernel

Architecture kernel module.

qemu

Provides the ability to exit QEMU and return a value as debug result.

timer

The timer support.

trap

Handles trap.

Macros

if_tdx_enabled

Inserts a TDX-specific code block.

Functions

read_random

Reads a hardware generated 64-bit random value.

read_tsc

Reads the current value of the processor’s time-stamp counter (TSC).

tsc_freq

Returns the frequency of TSC. The unit is Hz.