ostd::arch::device

Module cmos

Provides CMOS I/O port access.

“CMOS” is a tiny bit of very low power static memory that lives on the same chip as the Real-Time Clock (RTC).

Reference: https://wiki.osdev.org/CMOS

Statics

CMOS_ADDRESS

CMOS address I/O port

CMOS_DATA

CMOS data I/O port

Functions

century_register

Gets the century register location. This function is used in RTC(Real Time Clock) module initialization.