ostd

Attribute Macro early_cmdline_parser

#[early_cmdline_parser]

A macro attribute to register a parser for early command line arguments.

The attributed function will be used to parse the early command line arguments. The function takes a single parameter of type &str and returns an EarlyCmdline struct.

Note that the function will be called during an early boot stage. It must not cause soundness issues even if OSTD has not been fully initialized (e.g., the heap may be unavailable). OSDK enforces this by only allowing the function to be const.