ostd

Module boot

The architecture-independent boot module, which provides

a universal information getter interface from the bootloader to the rest of OSTD;

the routine booting into the actual kernel;

the routine booting the other processors in the SMP context.

Modules

memory_region

Information of memory regions in the boot phase.

smp

Symmetric multiprocessing (SMP) boot support.

Structs

BootInfo

The boot information provided by the bootloader.

BootloaderFramebufferArg

The framebuffer arguments.

Enums

BootloaderAcpiArg

ACPI information from the bootloader.

Functions

boot_info

Gets the boot information.