ostd::cpu::context::cpuid

Struct CpuId

pub struct CpuId { /* private fields */ }

The main type used to query information about the CPU weโre running on.

Other structs can be accessed by going through this type.

Implementations

impl CpuId

pub fn new() -> CpuId

Return new CpuId struct.

pub fn with_cpuid_fn(cpuid_fn: fn(_: u32, _: u32) -> CpuIdResult) -> CpuId

Return new CpuId struct with custom reader function.

This is useful for example when testing code or if we want to interpose on the CPUID calls this library makes.

pub fn get_vendor_info(&self) -> Option<VendorInfo>

Return information about the vendor (LEAF=0x00).

This leaf will contain a ASCII readable string such as โGenuineIntelโ for Intel CPUs or โAuthenticAMDโ for AMD CPUs.

Platforms

โ AMD โ Intel

pub fn get_feature_info(&self) -> Option<FeatureInfo>

Query a set of features that are available on this CPU (LEAF=0x01).

Platforms

โ AMD โ Intel

pub fn get_cache_info(&self) -> Option<CacheInfoIter>

Query basic information about caches (LEAF=0x02).

Platforms

โ AMD โ Intel

pub fn get_processor_serial(&self) -> Option<ProcessorSerial>

Retrieve serial number of processor (LEAF=0x03).

Platforms

โ AMD โ Intel

pub fn get_cache_parameters(&self) -> Option<CacheParametersIter>

Retrieve more elaborate information about caches (LEAF=0x04 or 0x8000_001D).

As opposed to get_cache_info, this will tell us about associativity, set size, line size of each level in the cache hierarchy.

Platforms

๐ก AMD โ Intel

pub fn get_monitor_mwait_info(&self) -> Option<MonitorMwaitInfo>

Information about how monitor/mwait works on this CPU (LEAF=0x05).

Platforms

๐ก AMD โ Intel

pub fn get_thermal_power_info(&self) -> Option<ThermalPowerInfo>

Query information about thermal and power management features of the CPU (LEAF=0x06).

Platforms

๐ก AMD โ Intel

pub fn get_extended_feature_info(&self) -> Option<ExtendedFeatures>

Find out about more features supported by this CPU (LEAF=0x07).

Platforms

๐ก AMD โ Intel

pub fn get_direct_cache_access_info(&self) -> Option<DirectCacheAccessInfo>

Direct cache access info (LEAF=0x09).

Platforms

โ AMD โ Intel

pub fn get_performance_monitoring_info( &self, ) -> Option<PerformanceMonitoringInfo>

Info about performance monitoring (LEAF=0x0A).

Platforms

โ AMD โ Intel

pub fn get_extended_topology_info(&self) -> Option<ExtendedTopologyIter>

Information about topology (LEAF=0x0B).

Intel SDM suggests software should check support for leaf 0x1F (CpuId::get_extended_topology_info_v2), and if supported, enumerate that leaf instead.

Platforms

โ AMD โ Intel

pub fn get_extended_topology_info_v2(&self) -> Option<ExtendedTopologyIter>

Extended information about topology (LEAF=0x1F).

Platforms

โ AMD โ Intel

pub fn get_extended_state_info(&self) -> Option<ExtendedStateInfo>

Information for saving/restoring extended register state (LEAF=0x0D).

Platforms

โ AMD โ Intel

pub fn get_rdt_monitoring_info(&self) -> Option<RdtMonitoringInfo>

Quality of service monitoring information (LEAF=0x0F).

Platforms

โ AMD โ Intel

pub fn get_rdt_allocation_info(&self) -> Option<RdtAllocationInfo>

Quality of service enforcement information (LEAF=0x10).

Platforms

โ AMD โ Intel

pub fn get_sgx_info(&self) -> Option<SgxInfo>

Information about secure enclave support (LEAF=0x12).

Platforms

โ AMD โ Intel

pub fn get_processor_trace_info(&self) -> Option<ProcessorTraceInfo>

Intel Processor Trace Enumeration Information (LEAF=0x14).

Platforms

โ AMD โ Intel

pub fn get_tsc_info(&self) -> Option<TscInfo>

Time Stamp Counter/Core Crystal Clock Information (LEAF=0x15).

Platforms

โ AMD โ Intel

pub fn get_processor_frequency_info(&self) -> Option<ProcessorFrequencyInfo>

Processor Frequency Information (LEAF=0x16).

Platforms

โ AMD โ Intel

pub fn get_soc_vendor_info(&self) -> Option<SoCVendorInfo>

Contains SoC vendor specific information (LEAF=0x17).

Platforms

โ AMD โ Intel

pub fn get_deterministic_address_translation_info(&self) -> Option<DatIter>

Query deterministic address translation feature (LEAF=0x18).

Platforms

โ AMD โ Intel

pub fn get_hypervisor_info(&self) -> Option<HypervisorInfo>

Returns information provided by the hypervisor, if running in a virtual environment (LEAF=0x4000_00xx).

Platform

Needs to be a virtual CPU to be supported.

pub fn get_extended_processor_and_feature_identifiers( &self, ) -> Option<ExtendedProcessorFeatureIdentifiers>

Extended Processor and Processor Feature Identifiers (LEAF=0x8000_0001).

Platforms

โ AMD ๐ก Intel

pub fn get_processor_brand_string(&self) -> Option<ProcessorBrandString>

Retrieve processor brand string (LEAF=0x8000_000{2..4}).

Platforms

โ AMD โ Intel

pub fn get_l1_cache_and_tlb_info(&self) -> Option<L1CacheTlbInfo>

L1 Instruction Cache Information (LEAF=0x8000_0005)

Platforms

โ AMD โ Intel (reserved)

pub fn get_l2_l3_cache_and_tlb_info(&self) -> Option<L2And3CacheTlbInfo>

L2/L3 Cache and TLB Information (LEAF=0x8000_0006).

Platforms

โ AMD ๐ก Intel

pub fn get_advanced_power_mgmt_info(&self) -> Option<ApmInfo>

Advanced Power Management Information (LEAF=0x8000_0007).

Platforms

โ AMD ๐ก Intel

pub fn get_processor_capacity_feature_info( &self, ) -> Option<ProcessorCapacityAndFeatureInfo>

Processor Capacity Parameters and Extended Feature Identification (LEAF=0x8000_0008).

Platforms

โ AMD ๐ก Intel

pub fn get_svm_info(&self) -> Option<SvmFeatures>

This function provides information about the SVM features that the processory supports. (LEAF=0x8000_000A)

If SVM is not supported if ExtendedProcessorFeatureIdentifiers::has_svm is false, this function is reserved then.

Platforms

โ AMD โ Intel

pub fn get_tlb_1gb_page_info(&self) -> Option<Tlb1gbPageInfo>

TLB 1-GiB Pages Information (LEAF=0x8000_0019)

Platforms

โ AMD โ Intel

pub fn get_performance_optimization_info( &self, ) -> Option<PerformanceOptimizationInfo>

Informations about performance optimization (LEAF=0x8000_001A)

Platforms

โ AMD โ Intel (reserved)

pub fn get_processor_topology_info(&self) -> Option<ProcessorTopologyInfo>

Informations about processor topology (LEAF=0x8000_001E)

Platforms

โ AMD โ Intel (reserved)

pub fn get_memory_encryption_info(&self) -> Option<MemoryEncryptionInfo>

Informations about memory encryption support (LEAF=0x8000_001F)

Platforms

โ AMD โ Intel (reserved)