ostd::cpu::local

Struct CpuLocalCell

pub struct CpuLocalCell<T: 'static>(/* private fields */);

Inner mutable CPU-local objects.

CPU-local cell objects are only accessible from the current CPU. When accessing an underlying object using the same CpuLocalCell instance, the actually accessed object is always on the current CPU. So in a preemptive kernel task, the operated object may change if interrupts are enabled.

The inner mutability is provided by single instruction operations, and the CPU-local cell objects will not ever be shared between CPUs. So it is safe to modify the inner value without any locks.

You should only create the CPU-local cell object using the macro cpu_local_cell!.

Please exercise extreme caution when using CpuLocalCell. In most cases, it is necessary to disable interrupts or preemption when using it to prevent the operated object from being changed, which can lead to race conditions.

For the difference between super::CpuLocal and CpuLocalCell, see super.

Implementations

impl<T: 'static> CpuLocalCell<T>

pub fn as_mut_ptr(&'static self) -> *mut T

Gets access to the underlying value through a raw pointer.

This function calculates the virtual address of the CPU-local object based on the CPU-local base address and the offset in the BSP.

This method is safe, but using the returned pointer will be unsafe. Specifically,

Preemption should be disabled from the time this method is called to the time the pointer is used. Otherwise, the pointer may point to the variable on another CPU, making it difficult or impossible to determine if the data can be borrowed.

If the variable can be used in interrupt handlers, borrowing the data should be done with interrupts disabled. Otherwise, more care must be taken to ensure that the borrowing rules are correctly enforced, since the interrupts may come asynchronously.

impl<T: 'static + SingleInstructionAddAssign<T>> CpuLocalCell<T>

pub fn add_assign(&'static self, rhs: T)

Adds a value to the per-CPU object in a single instruction.

This operation wraps on overflow/underflow.

Note that this memory operation will not be elided or reordered by the compiler since it is a black-box.

impl<T: 'static + SingleInstructionSubAssign<T>> CpuLocalCell<T>

pub fn sub_assign(&'static self, rhs: T)

Subtracts a value to the per-CPU object in a single instruction.

This operation wraps on overflow/underflow.

Note that this memory operation will not be elided or reordered by the compiler since it is a black-box.

impl<T: 'static + SingleInstructionBitAndAssign<T>> CpuLocalCell<T>

pub fn bitand_assign(&'static self, rhs: T)

Bitwise ANDs a value to the per-CPU object in a single instruction.

Note that this memory operation will not be elided or reordered by the compiler since it is a black-box.

impl<T: 'static + SingleInstructionBitOrAssign<T>> CpuLocalCell<T>

pub fn bitor_assign(&'static self, rhs: T)

Bitwise ORs a value to the per-CPU object in a single instruction.

Note that this memory operation will not be elided or reordered by the compiler since it is a black-box.

impl<T: 'static + SingleInstructionBitXorAssign<T>> CpuLocalCell<T>

pub fn bitxor_assign(&'static self, rhs: T)

Bitwise XORs a value to the per-CPU object in a single instruction.

Note that this memory operation will not be elided or reordered by the compiler since it is a black-box.

impl<T: 'static + SingleInstructionLoad> CpuLocalCell<T>

pub fn load(&'static self) -> T

Gets the value of the per-CPU object in a single instruction.

Note that this memory operation will not be elided or reordered by the compiler since it is a black-box.

impl<T: 'static + SingleInstructionStore> CpuLocalCell<T>

pub fn store(&'static self, val: T)

Writes a value to the per-CPU object in a single instruction.

Note that this memory operation will not be elided or reordered by the compiler since it is a black-box.