ostd

Derive Macro Pod

#[derive(Pod)]

Deriving Pod trait for a struct or union.

When deriving the Pod trait, this macro performs a safety check because the Pod trait is marked as unsafe. For structs and unions, the macro checks that the struct has a valid repr attribute (e.g., repr(C), repr(u8)), and each field is of Pod type. Enums cannot implement the Pod trait.

If you want to implement Pod for a struct or union with fields that are not of Pod type, you can implement it unsafely and perform the necessary checks manually.