ostd::input

Enum KeySym

#[repr(u32)]
pub enum KeySym {

    Unknown = 0,
    Escape = 1,
    Return = 2,
    Backspace = 3,
    Tab = 4,
    Delete = 5,
    Insert = 6,
    Up = 16,
    Down = 17,
    Left = 18,
    Right = 19,
    Home = 32,
    End = 33,
    PageUp = 34,
    PageDown = 35,
    F1 = 257,
    F2 = 258,
    F3 = 259,
    F4 = 260,
    F5 = 261,
    F6 = 262,
    F7 = 263,
    F8 = 264,
    F9 = 265,
    F10 = 266,
    F11 = 267,
    F12 = 268,
    Printable = 32_768,

}

Virtual key identifier — layout-independent.

Printable characters are represented by their Unicode code point in the range 0x0020..=0xFFFF. Special keys use the negative range or values above 0x1_0000.

Variants

Unknown = 0

Escape = 1

Return = 2

Backspace = 3

Tab = 4

Delete = 5

Insert = 6

Up = 16

Down = 17

Left = 18

Right = 19

Home = 32

End = 33

PageUp = 34

PageDown = 35

F1 = 257

F2 = 258

F3 = 259

F4 = 260

F5 = 261

F6 = 262

F7 = 263

F8 = 264

F9 = 265

F10 = 266

F11 = 267

F12 = 268

Printable = 32_768