ostd::irq

Function register_bottom_half_handler_l1

pub fn register_bottom_half_handler_l1(
    func: fn(_: DisabledLocalIrqGuard) -> DisabledLocalIrqGuard,
)

Registers a bottom half callback to be executed at interrupt level 1.

The callback takes a DisabledLocalIrqGuard as an argument. This allows the callback to drop the guard in order to re-enable IRQs on the current CPU. The callback requires returning a DisabledLocalIrqGuard, thus ensuring that local IRQs are disabled by the end of the callback.

The function may be called only once; subsequent calls take no effect.