Module ostd::mm

Virtual memory (VM).

Re-exports

pub use self::frame::options::FrameAllocOptions;

pub use self::frame::Frame;

pub use self::frame::FrameVec;

pub use self::frame::FrameVecIter;

pub use self::frame::Segment;

Modules

frame

Untyped physical memory management.

Structs

DmaCoherent

A coherent (or consistent) DMA mapping, which guarantees that the device and the CPU can access the data in parallel.

DmaStream

A streaming DMA mapping. Users must synchronize data before reading or after writing to ensure consistency.

DmaStreamSlice

A slice of streaming DMA mapping.

KernelSpace

A marker structure used for VmReader and VmWriter, representing their operated memory scope is in kernel space.

PageFlags

Page protection permissions and access status.

PageProperty

The property of a mapped virtual memory page.

UserSpace

A marker structure used for VmReader and VmWriter, representing their operated memory scope is in user space.

VmMapOptions

Options for mapping physical memory pages into a VM address space. See VmSpace::map.

VmReader

VmReader is a reader for reading data from a contiguous range of memory.

VmSpace

Virtual memory space.

VmWriter

VmWriter is a writer for writing data to a contiguous range of memory.

Enums

CachePolicy

A type to control the cacheability of the main memory.

DmaDirection

DmaDirection limits the data flow direction of DmaStream and prevents users from reading and writing to DmaStream unexpectedly.

Constants

MAX_USERSPACE_VADDR

The maximum virtual address of user space (non inclusive).

PAGE_SIZE

The page size

Statics

FRAMEBUFFER_REGIONS

Memory regions used for frame buffer.

Traits

HasDaddr

A trait for types that have mapped address in the device address space.

HasPaddr

Gets physical address trait

VmIo

A trait that enables reading/writing data from/to a VM object, e.g., VmSpace, FrameVec, and Frame.

Functions

is_page_aligned

Checks if the given address is page-aligned.

Type Aliases

Daddr

The devide address.

Paddr

Physical addresses.

PagingLevel

The level of a page table node or a frame.

Vaddr

Virtual addresses.