ostd::mm

Module dma

Direct Memory Access (DMA).

This module provides DmaCoherent and DmaStream abstractions for managing DMA memory regions with different remapping, caching and synchronization requirements.

Structs

DmaCoherent

A DMA memory object that can be accessed in a cache-coherent manner.

DmaStream

A DMA memory object with streaming access.

Enums

FromAndToDevice

Data flows both from and to the device.

FromDevice

Data flows from the device.

ToDevice

Data flows to the device.

Traits

DmaDirection

DmaDirection limits the data flow direction of DmaStream and prevents users from reading and writing to DmaStream unexpectedly.