ostd::mm::dma

Struct DmaStream

pub struct DmaStream<D: DmaDirection = FromAndToDevice> { /* private fields */ }

A DMA memory object with streaming access.

The kernel must synchronize the data by sync_from_device/sync_to_device when interacting with the device.

Implementations

impl<D: DmaDirection> DmaStream<D>

pub fn alloc(nframes: usize, is_cache_coherent: bool) -> Result<Self, Error>

Allocates a region of physical memory for streaming DMA access.

The memory of the newly-allocated DMA buffer is initialized to zeros. This method is only available when D is ToDevice or FromAndToDevice, as zeroing requires write access to the buffer.

The is_cache_coherent argument specifies whether the target device that the DMA mapping is prepared for can access the main memory in a CPU cache coherent way or not.

Comparison with DmaStream::map

This method is semantically equivalent to allocating a USegment via FrameAllocOptions::alloc_segment and then mapping it with DmaStream::map. However, DmaStream::alloc combines these two operations and can be more efficient in certain scenarios, particularly in confidential VMs, where the overhead of bounce buffers can be avoided.

pub fn alloc_uninit( nframes: usize, is_cache_coherent: bool, ) -> Result<Self, Error>

Allocates a region of physical memory for streaming DMA access without initialization.

This method is the same as DmaStream::alloc except that it skips zeroing the memory of newly-allocated DMA region.

pub fn map(segment: USegment, is_cache_coherent: bool) -> Result<Self, Error>

Establishes DMA stream mapping for a given USegment.

The is_cache_coherent argument specifies whether the target device that the DMA mapping is prepared for can access the main memory in a CPU cache coherent way or not.

pub fn sync_from_device(&self, byte_range: Range<usize>) -> Result<(), Error>

Synchronizes the streaming DMA mapping data from the device.

This method should be called when the data of the streaming DMA mapping has been updated by the device side. Before the CPU side starts to read (e.g., using read_bytes), it must call the Self::sync_from_device method first.

pub fn sync_to_device(&self, byte_range: Range<usize>) -> Result<(), Error>

Synchronizes the streaming DMA mapping data to the device.

This method should be called when the data of the streaming DMA mapping has been updated by the CPU side (e.g., using write_bytes). Before the CPU side notifies the device side to read, it must call the Self::sync_to_device method first.