ostd::mm

Module fault

Page fault handling.

This module manages a handler that can address page faults occurring in the kernel due to user-space addresses. For example, this occurs during FallibleVmRead::read_fallible and FallibleVmWrite::write_fallible.

If the page fault is handled successfully, the read/write process continues and the fault address is retried. Otherwise, those methods will return an error to indicate that the read/write process cannot be completed.

Functions

inject_user_page_fault_handler

Injects a custom handler for page faults that occur in the kernel and are caused by user-space addresses.

Type Aliases

UserPageFaultHandler

A handler that handles page faults caused by user-space addresses.