Module ostd::mm::frame

Untyped physical memory management.

A frame is a special page that is untyped memory. It is used to store data irrelevant to the integrity of the kernel. All pages mapped to the virtual address space of the users are backed by frames. Frames, with all the properties of pages, can additionally be safely read and written by the kernel or the user.

Re-exports

pub use frame_vec::FrameVec;

pub use frame_vec::FrameVecIter;

pub use segment::Segment;

Modules

frame_vec

Page frames.

options

Options for allocating frames

segment

A contiguous range of page frames.

Structs

Frame

A handle to a physical memory page of untyped memory.

FrameRef

FrameRef is a struct that can work as &'a Frame.