ostd::mm::frame::segment

Type Alias USegment

pub type USegment = Segment<dyn AnyUFrameMeta>;

A contiguous range of homogeneous untyped physical memory frames that have any metadata.

In other words, the metadata of the frames are of the same type, and they are untyped, but the type of metadata is not known at compile time. An USegment as a parameter accepts any untyped segments.

The usage of this frame will not be changed while this object is alive.

Aliased Type

pub struct USegment { /* private fields */ }