ostd::mm::frame::untyped

Trait UntypedMem

pub trait UntypedMem {
    // Required methods
    fn reader(&self) -> VmReader<'_, Infallible>;
    fn writer(&self) -> VmWriter<'_, Infallible>;
}

A physical memory range that is untyped.

Untyped frames or segments can be safely read and written by the kernel or the user.

Required Methods

fn reader(&self) -> VmReader<'_, Infallible>

Borrows a reader that can read the untyped memory.

fn writer(&self) -> VmWriter<'_, Infallible>

Borrows a writer that can write the untyped memory.