ostd::mm::io

Trait PodOnce

pub trait PodOnce: Pod { }

A marker trait for POD types that can be read or written with one instruction.

This trait is mostly a hint, since it’s safe and can be implemented for any POD type. If it is implemented for a type that cannot be read or written with a single instruction, calling read_once/write_once will lead to a failed compile-time assertion.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.