ostd::mm

Module io_util

Utilities for types in super::io.

Enums

VmReaderWriterIdentity

A marker type that denotes reader and writer identities as HasVmReaderWriter return types.

VmReaderWriterResult

A marker type that denotes reader and writer results as HasVmReaderWriter return types.

Traits

HasVmReaderWriter

A helper trait that denotes types that can provide VmReaders and VmWriters.

VmReaderWriterTypes

A marker trait that denotes the return types for HasVmReaderWriter.