ostd::mm::io_util

Trait HasVmReaderWriter

pub trait HasVmReaderWriter {
    type Types: VmReaderWriterTypes;

    // Required methods
    fn reader(&self) -> <Self::Types as VmReaderWriterTypes>::Reader<'_>;
    fn writer(&self) -> <Self::Types as VmReaderWriterTypes>::Writer<'_>;
}

A helper trait that denotes types that can provide VmReaders and VmWriters.

Having the reader and writer means that the type is capable of performing a range of VM operations. Thus, several traits will be automatically and efficiently implemented, such as VmIo, VmIoFill, and VmIoOnce.

Required Associated Types

type Types: VmReaderWriterTypes

A marker type that denotes the return types of Self::reader and Self::writer.

This can be either VmReaderWriterIdentity or VmReaderWriterResult.

Required Methods

fn reader(&self) -> <Self::Types as VmReaderWriterTypes>::Reader<'_>

Returns a reader to read data from it.

fn writer(&self) -> <Self::Types as VmReaderWriterTypes>::Writer<'_>

Returns a writer to write data to it.