ostd::mm

Trait HasDaddr

pub trait HasDaddr {
    // Required method
    fn daddr(&self) -> Daddr;
}

Memory objects that have a mapped address in the device address space.

Required Methods

fn daddr(&self) -> Daddr

Returns the base address of the mapping in the device address space.