ostd

Module power

Power management.

Enums

ExitCode

An exit code that denotes the reason for restarting or powering off.

Functions

inject_poweroff_handler

Injects a handler that can power off the system.

inject_restart_handler

Injects a handler that can restart the system.

poweroff

Powers off the system.

restart

Restarts the system.