ostd::smp

Function inter_processor_call

pub fn inter_processor_call(targets: &CpuSet, f: fn())

Executes a function on other processors.

The provided function f will be executed on all target processors specified by targets. It can also be executed on the current processor. The function should be short and non-blocking, as it will be executed in interrupt context with interrupts disabled.

This function does not block until all the target processors acknowledges the interrupt. So if any of the target processors disables IRQs for too long that the controller cannot queue them, the function will not be executed.

The function f will be executed asynchronously on the target processors. However if called on the current processor, it will be synchronous.