ostd::task::scheduler

Function inject_scheduler

pub fn inject_scheduler(scheduler: &'static dyn Scheduler<Task>)

Injects a custom implementation of task scheduler into OSTD.

This function can only be called once and must be called during the initialization phase of kernel, before any Task-related APIs are invoked.