ostd::task

Struct CurrentTask

pub struct CurrentTask(/* private fields */);

The current task.

This type is not Send, so it cannot outlive the current task.

This type is also not Sync, so it can provide access to the local data of the current task.

Implementations

impl CurrentTask

pub fn local_data(&self) -> &(dyn Any + Send)

Returns the local data of the current task.

Note that the local data is only accessible in the task context. Although there is a current task in the non-task context (e.g. IRQ handlers), access to the local data is forbidden as it may cause soundness problems.

Panics

This method will panic if called in a non-task context.

pub fn cloned(&self) -> Arc<Task>

Returns a cloned Arc<Task>.