ostd::trap

Function register_bottom_half_handler

pub fn register_bottom_half_handler(
    func: fn(_: DisabledLocalIrqGuard) -> DisabledLocalIrqGuard,
)

Registers a function to the interrupt bottom half execution.

The handling of an interrupt can be divided into two parts: top half and bottom half. Top half typically performs critical tasks and runs at a high priority. Relatively, bottom half defers less critical tasks to reduce the time spent in hardware interrupt context, thus allowing the interrupts to be handled more quickly.

The bottom half handler is called following the execution of the top half. Because the handler accepts a DisabledLocalIrqGuard as a parameter, interrupts are still disabled upon entering the handler. However, the handler can enable interrupts by internally dropping the guard. When the handler returns, interrupts should remain disabled, as the handler is expected to return an IRQ guard.

This function can only be registered once. Subsequent calls will do nothing.