ostd::util::id_set

Trait IdSetSlicer

pub trait IdSetSlicer<I: Id> {
    // Required method
    fn to_range_bounds(self) -> (usize, usize);
}

A trait that unifies all types that slice a portion of IdSet.

Required Methods

fn to_range_bounds(self) -> (usize, usize)

Converts the index type to inclusive start and exclusive end bounds.

Returns (start, end) where:

start: inclusive lower bound

end: exclusive upper bound