Crate ostd

The standard library for Asterinas and other Rust OSes.

Re-exports

pub use self::cpu::CpuLocal;

pub use self::prelude::Result;

Modules

arch

Platform-specific code.

boot

The architecture-independent boot module, which provides a universal interface from the bootloader to the rest of OSTD.

bus

Bus operations

collections

This module provides some advanced collections.

console

Console output.

cpu

CPU.

io_mem

I/O memory.

logger

Logging support.

mm

Virtual memory (VM).

panicking

Panic support.

prelude

The prelude.

sync

Useful synchronization primitives.

task

Tasks are the unit of code execution.

trap

Handles trap across kernel and user space.

user

User space.

Macros

cpu_local

Defines a CPU-local variable.

early_print

Prints to the console.

early_println

Prints to the console, with a newline.

offset_of

Gets the offset of a field within a type as a pointer.

value_offset

Gets the offset of a field within an object as a pointer.

Enums

Error

The error type which is returned from the APIs of this crate.

Functions

init

Initializes OSTD.

Attribute Macros

main

This macro is used to mark the kernel entry point.