ostd::mm

Module vm_space

Virtual memory space management.

The VmSpace struct is provided to manage the virtual memory space of a user. Cursors are used to traverse and modify over the virtual memory space concurrently. The VM space cursor self::Cursor is just a wrapper over the page table cursor, providing efficient, powerful concurrent accesses to the page table.

Structs

Cursor

The cursor for querying over the VM space without modifying it.

CursorMut

The cursor for modifying the mappings in VM space.

VmSpace

A virtual address space for user-mode tasks, enabling safe manipulation of user-space memory.

Enums

VmQueriedItem

The result of a query over the VM space.