Module ostd::task

Tasks are the unit of code execution.

Structs

DisablePreemptGuard

A guard for disable preempt.

FifoScheduler

A simple FIFO (First-In-First-Out) task scheduler.

Priority

The priority of a task.

Task

A task that executes a function to the end.

TaskAdapter

TaskOptions

Options to create or spawn a new task.

Enums

TaskStatus

The status of a task.

Traits

Scheduler

A scheduler for tasks.

TaskContextApi

Trait for manipulating the task context.

Functions

add_task

Adds a task to the global scheduler.

current_task

Retrieves the current task running on the processor.

disable_preempt

Disables preemption.

preempt

Preempts the task.

schedule

Calls this function to switch to other task by using GLOBAL_SCHEDULER

set_scheduler

Sets the global task scheduler.